+++ src/config/api.js
@@ -1,0 +1,8 @@
+const API_CONFIG = {
+    baseUrl: '/v2',
+    headers: {
+        "Content-Type": "application/json"
+    }
+};
+
+export default API_CONFIG;
